#Issue 2  - using styled components on Input
