+++ src/components/index.ts
@@ -4,7 +4,7 @@ export { CreateTask } from './createTask'
 export { DeleteTaskButton } from './deleteTaskButton/deleteTaskButton'
 */
 export { Header } from './header'
-// export { Input } from './input/input'
+export { Input } from './input'
 export { Logo } from './logo/logo'
 /*
 export { NoTasks } from './noTasks/noTasks'
