--- src/components/input/input.module.scss
@@ -1,22 +1,0 @@
-@import '@assets/colors';
-
-.input {
-  flex: 1;
-  border: none;
-  align-self: flex-end;
-
-  padding: 1rem;
-  height: 3.375rem;
-  border-radius: 5px;
-
-  color: $gray-100;
-  background-color: $gray-500;
-
-  &::placeholder {
-    color: $gray-300;
-  }
-
-  &:focus {
-    //border: 1px solid $purple-dark;
-  }
-}
